*** empty log message ***
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@97292 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 12bc6d302f8d930469554ad3aad1cefc9fc5ebdf

+++ Online/Online/MBMDump/src/BankWindow.cpp
@@ -32,11 +32,11 @@ BankWindow::BankWindow(BaseMenu* par,int cmd_id, const Format& f, const RawBank*
       memset(txt,' ',sizeof(txt));
       // 0=nothing,1=offset,2=line number
       if(m_fmt.column_one_flag == 1)  {
-        sprintf(tmp,"%4ld",(char*)data-(char*)b->data());
+        sprintf(tmp,"%4d",(int)((char*)data-(char*)b->data()));
         memcpy(txt+3,tmp,4);
       }
       else if(m_fmt.column_one_flag == 2)  {
-        sprintf(tmp,"%4ld",(data-b->data())/m_fmt.words_per_line);
+        sprintf(tmp,"%4d",(int)((data-b->data())/m_fmt.words_per_line));
         memcpy(txt+3,tmp,4);
       }
       else  {
